+++ tests/src/core/importType.js
@@ -1,7 +1,7 @@
 import { expect } from 'chai'
 import * as path from 'path'
 
-import importType, {isExternalModule} from 'core/importType'
+import importType, {isExternalModule, isScopedModule} from 'core/importType'
 
 import { testContext, testFilePath } from '../utils'
 
@@ -236,5 +236,10 @@ describe('importType(name)', function () {
     expect(isExternalModule('foo', {
       'import/external-module-folders': ['E:\\path\\to\\node_modules'],
     }, 'E:\\path\\to\\node_modules\\foo')).to.equal(true)
+  })
+
+  it('correctly identifies scoped modules with `isScopedModule`', () => {
+    expect(isScopedModule('@/abc')).to.equal(false)
+    expect(isScopedModule('@a/abc')).to.equal(true)
   })
 })