+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This change log adheres to standards from [Keep a CHANGELOG](http://keepachangel
 
 ### Fixed
 - [`default`]/TypeScript: avoid crash on `export =` with a MemberExpression ([#1841], thanks [@ljharb])
+- [`extensions`]/importType: Fix @/abc being treated as scoped module ([#1854], thanks [@3nuc])
 
 ## [2.22.0] - 2020-06-26
 ### Added
@@ -725,6 +726,7 @@ for info on changes for earlier releases.
 
 [`memo-parser`]: ./memo-parser/README.md
 
+[#1854]: https://github.com/benmosher/eslint-plugin-import/issues/1854
 [#1841]: https://github.com/benmosher/eslint-plugin-import/issues/1841
 [#1836]: https://github.com/benmosher/eslint-plugin-import/pull/1836
 [#1835]: https://github.com/benmosher/eslint-plugin-import/pull/1835
@@ -1260,3 +1262,4 @@ for info on changes for earlier releases.
 [@be5invis]: https://github.com/be5invis
 [@noelebrun]: https://github.com/noelebrun
 [@beatrizrezener]: https://github.com/beatrizrezener
+[@3nuc]: https://github.com/3nuc